Split pi-codex websocket internals
